+++ addon/components/flexberry-lookup.js
@@ -8,7 +8,6 @@ import FlexberryBaseComponent from './flexberry-base-component';
 import { translationMacro as t } from 'ember-i18n';
 import { getRelationType } from 'ember-flexberry-data/utils/model-functions';
 import { Query } from 'ember-flexberry-data';
-import deserializeSortingParam from '../utils/deserialize-sorting-param';
 
 const {
   Builder,
@@ -199,15 +198,6 @@ export default FlexberryBaseComponent.extend({
   sorting: 'asc',
 
   /**
-    Ordering condition for list of records to choose.
-    Expected string type: '+Name1-Name2...', where: '+' and '-' - sorting direction, 'NameX' - property name for soring.
-
-    @property orderBy
-    @type String
-  */
-  orderBy: undefined,
-
-  /**
     Classes by property of autocomplete.
 
     @property autocompleteClass
@@ -229,7 +219,7 @@ export default FlexberryBaseComponent.extend({
   */
   folvComponentName: Ember.computed('componentName', function() {
     let componentName = this.get('componentName') || 'undefined';
-    return `folv_in_${componentName}_lookup`;
+    return `${componentName}`;
   }),
 
   /**
@@ -340,9 +330,7 @@ export default FlexberryBaseComponent.extend({
     'lookupLimitPredicate',
     'relatedModel',
     '_lookupWindowCustomPropertiesData',
-    'orderBy',
     function() {
-      let ordering = this.get('orderBy') ? this.get('orderBy') : '';
       let perPage = this.get('userSettings').getCurrentPerPage(this.get('folvComponentName'));
       return {
         projection: this.get('projection'),
@@ -354,7 +342,6 @@ export default FlexberryBaseComponent.extend({
         componentName: this.get('componentName'),
         notUseUserSettings: this.get('notUseUserSettings'),
         perPage: perPage || this.get('perPage'),
-        sorting: deserializeSortingParam(ordering),
         folvComponentName: this.get('folvComponentName'),
 
         //TODO: move to modal settings.